consumer/requirements.txt
# Lists the Python dependencies for the consumer.
kafka-python
psycopg2-binary